+++ biosimulator_processes/data_model/compare_data_model.py
@@ -3,8 +3,41 @@ from abc import ABC
 from dataclasses import dataclass
 from biosimulator_processes.utils import prepare_single_ode_process_document
 
+from typing import List, Dict, Tuple, Any
+from biosimulator_processes.data_model import _BaseModel as BaseModel
 
-__all__ = ['ComparisonDocument']
+
+class SimulatorComparisonResult(BaseModel):
+    simulators: List[str]
+    value: Dict[Tuple[str], Dict[str, Any]]
+
+
+class IntervalResult(BaseModel):
+    global_time_stamp: float
+    results: Dict[str, Any]
+
+
+class SimulatorResult(BaseModel):
+    process_id: str
+    simulator: str
+    result: List[IntervalResult]
+
+
+class ComparisonResults(BaseModel):
+    duration: int
+    num_steps: int
+    values: List[SimulatorResult]
+
+
+class ProcessAttributes(BaseModel):
+    name: str
+    initial_state: Dict[str, Any]
+    inputs: Dict[str, Any]
+    outputs: Dict[str, Any]
+
+
+class CompositeRunError(BaseModel):
+    exception: Exception
 
 
 class ComparisonDocument(ABC):
