+++ compare_api/main.py
@@ -5,7 +5,7 @@ from urllib.parse import unquote
 
 from fastapi import FastAPI, HTTPException, Query, File, UploadFile
 from fastapi.middleware.cors import CORSMiddleware
-from compare_api.datamodel import SimulatorComparisonResult, CompositeRunError, ProcessAttributes, ComparisonResults
+from biosimulator_processes.data_model.compare_data_model import ProcessAttributes, ComparisonResults
 from compare_api.src.composite_doc import create_comparison_document, generate_workflow, run_workflow
 
 
@@ -81,7 +81,7 @@ async def get_process_attributes(
 
 @app.post(
     "/run-comparison",
-    response_model=SimulatorComparisonResult,
+    response_model=ComparisonResults,
     name="Run a Simulator Comparison",
     operation_id="run-comparison",
     responses={
@@ -92,7 +92,7 @@ async def run_comparison(
         simulators: List[str] = Query(..., title="Simulators to Compare"),
         duration: int = Query(..., title="Duration"),
         num_steps: int = Query(..., title="Number of Steps")
-) -> SimulatorComparisonResult:
+) -> ComparisonResults:
     # TODO: Add fallback of biosimulations 1.0 for simulators not yet implemented.
     # TODO: enable remote model file for model path with download
     try:
@@ -101,7 +101,7 @@ async def run_comparison(
         document = create_comparison_document(model_file_location, simulators, duration, num_steps)
         workflow = generate_workflow(document)
         results = await run_workflow(workflow, duration)
-        return SimulatorComparisonResult(simulators=simulators, value=results)
+        return ComparisonResults(simulators=simulators, value=results)
     except AssertionError as e:
         logger.warning(f'failed to run simulator comparison composite: {str(e)}')
         raise HTTPException(status_code=404, detail="Parameters not valid.")
